Make AbstractCodeGenerator.apply(module) to be abstract interface
The 'apply' function aims to translate Whiley byte-code into C code. As
C and OpenCL have different code structure and code sequences, apply
function should be made protected and defined in the implementation.
